history - move commands to own file
